core/user/requirements.txt
flask==1.0.2
kafka-python==1.4.3
twisted==18.9.0
klein==17.10.0
pypiwin32
